# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangeRequestDetailOut": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ApprovalOut"
            },
            "title": "Approvals",
            "type": "array"
          },
          "base_entity_versions": {
            "additionalProperties": true,
            "title": "Base Entity Versions",
            "type": "object"
          },
          "before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "change_request": {
            "$ref": "#/components/schemas/ChangeRequestOut"
          },
          "evidence_refs": {
            "items": {},
            "title": "Evidence Refs",
            "type": "array"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/AuditEventOut"
            },
            "title": "History",
            "type": "array"
          },
          "impact": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/RequirementOut"
            },
            "title": "Requirements",
            "type": "array"
          },
          "runs": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Runs",
            "type": "array"
          },
          "viewer": {
            "$ref": "#/components/schemas/ReviewerOut"
          }
        },
        "required": [
          "change_request",
          "payload",
          "before",
          "after",
          "impact",
          "evidence_refs",
          "base_entity_versions",
          "requirements",
          "approvals",
          "viewer",
          "runs",
          "history"
        ],
        "title": "ChangeRequestDetailOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewIn": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "default": "",
            "title": "Comment",
            "type": "string"
          }
        },
        "title": "ReviewIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Relay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/change-requests/{change_id}": {
      "get": {
        "operationId": "get_change_request_api_v1_change_requests__change_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Change Request",
        "tags": [
          "governance"
        ]
      },
      "put": {
        "operationId": "update_change_request_api_v1_change_requests__change_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequestUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Change Request",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/change-requests/{change_id}/approve": {
      "post": {
        "description": "Record an approval. The response status is ``stale`` when the base changed meanwhile.",
        "operationId": "approve_change_request_api_v1_change_requests__change_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOutcomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Change Request",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/change-requests/{change_id}/reject": {
      "post": {
        "operationId": "reject_change_request_api_v1_change_requests__change_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOutcomeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Change Request",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/change-requests/{change_id}/submit": {
      "post": {
        "operationId": "submit_change_request_api_v1_change_requests__change_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Change Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JustificationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Change Request",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}